+++ @jambaree/next-wordpress/src/WordpressTemplate.tsx
@@ -1,11 +1,12 @@
-import React, { use } from "react";
-
+import React from "react";
+// import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import getSeedData from "./getSeedData";
-import getAllContentNodePaths from "./getAllContentNodePaths";
+// import getAllContentNodePaths from "./getAllContentNodePaths";
 import getTemplate from "./getTemplate";
+// import getYoastData from "./getYoastData";
 
-export default function WordpressTemplate({
+export default async function WordpressTemplate({
   paths,
   templates,
 }: {
@@ -13,34 +14,19 @@ export default function WordpressTemplate({
   templates: any;
 }) {
   const uri = paths?.join?.("/") || "/";
-  const seedData = use(getSeedData({ uri }));
+  const seedData = await getSeedData({ uri });
 
   if (!seedData) {
     notFound();
   }
 
-  const PageTemplate = use(
-    getTemplate({
-      seedData,
-      templates,
-    })
-  );
-
+  const PageTemplate = await getTemplate({
+    seedData,
+    templates,
+  });
   if (!PageTemplate) {
     notFound();
   }
 
   return <PageTemplate uri={uri} seedData={seedData} paths={paths} />;
-}
-
-export const revalidate = "force-cache";
-
-export async function generateStaticParams() {
-  const nodePaths = await getAllContentNodePaths();
-
-  return nodePaths.map((node) => {
-    return {
-      paths: [node.uri || "/"],
-    };
-  });
 }