+++ @jambaree/next-wordpress/CHANGE_LOG.md
@@ -15,3 +15,24 @@
   0.1.7
 
 - added YoastSEO support for taxonomies
+
+  0.1.8
+
+- Remove generateStaticParam for debugging
+
+  0.1.9
+
+- Add generateStaticParam
+
+  0.2.0
+
+- Changed getMenuItems - > useMenuItems and reduced complexity
+
+  0.2.1
+
+- Changed getMenuItems - > useMenuItems in index
+
+  0.2.2
+
+- Removed YoastSEO as there is a new way to handle this now with beta.next app directory (generateMetaData)
+- Remove generateStaticParams from package